+++ packages/calcite-components/src/custom-theme/swatch.ts
@@ -1,0 +1,7 @@
+import { html } from "../../support/formatting";
+
+export const swatchTokens = {
+  calciteSwatchCornerRadius: "",
+};
+
+export const swatch = html` <calcite-swatch color="#aabbcc" label="example" value="calcite swatch"></calcite-swatch> `;